# src/main/java/com/simbirsoft/my_app/api-doc.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"Electricity","description":"Electricity Controller"},{"name":"Expense","description":"Expense Controller"},{"name":"Rate","description":"Rate Controller"},{"name":"WaterSupply","description":"Water Supply Controller"},{"name":"basic-error-controller","description":"Basic Error Controller"}],"paths":{"/api/v1/electricity/create":{"post":{"tags":["Electricity"],"summary":"create new electricity","operationId":"addElectDataUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"electricityDto","description":"electricityDto","required":true,"schema":{"$ref":"#/definitions/ElectricityDto"}},{"name":"rateE","in":"query","required":false,"type":"number","format":"double"},{"name":"rateWC","in":"query","required":false,"type":"number","format":"double"},{"name":"rateWH","in":"query","required":false,"type":"number","format":"double"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/v1/electricity/{id}":{"get":{"tags":["Electricity"],"summary":"find by id","operationId":"getByIdUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Electricity"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["Electricity"],"summary":"delete electricity by id","operationId":"deleteUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/api/v1/expense/create":{"post":{"tags":["Expense"],"summary":"create new expense","operationId":"addElectDataUsingPOST_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"expenseDto","description":"expenseDto","required":true,"schema":{"$ref":"#/definitions/ExpenseDto"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/v1/expense/{id}":{"get":{"tags":["Expense"],"summary":"find rate by id","operationId":"getByIdUsingGET_1","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Expense"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["Expense"],"summary":"delete expense by id","operationId":"deleteUsingDELETE_1","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/api/v1/rate/create":{"post":{"tags":["Rate"],"summary":"create new rate","operationId":"addElectDataUsingPOST_2","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"rateDto","description":"rateDto","required":true,"schema":{"$ref":"#/definitions/RateDto"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/v1/rate/{id}":{"get":{"tags":["Rate"],"summary":"Get rate information","operationId":"getByIdUsingGET_2","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"Rate id","required":true,"type":"integer","format":"int64","x-example":1}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Rate"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["Rate"],"summary":"delete old rate","operationId":"deleteUsingDELETE_2","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/api/v1/waterSupply/create":{"post":{"tags":["WaterSupply"],"summary":"create new water supply","operationId":"addElectDataUsingPOST_3","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"rateE","in":"query","required":false,"type":"number","format":"double"},{"name":"rateWC","in":"query","required":false,"type":"number","format":"double"},{"name":"rateWH","in":"query","required":false,"type":"number","format":"double"},{"in":"body","name":"waterSupplyDto","description":"waterSupplyDto","required":true,"schema":{"$ref":"#/definitions/WaterSupplyDto"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/v1/waterSupply/{id}":{"get":{"tags":["WaterSupply"],"summary":"find by id","operationId":"getByIdUsingGET_3","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/WaterSupply"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["WaterSupply"],"summary":"delete water supply by id","operationId":"deleteUsingDELETE_3","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/error":{"get":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingGET","produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"head":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingHEAD","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"post":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingPOST","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingPUT","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingDELETE","produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"options":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingOPTIONS","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"patch":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingPATCH","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}}},"definitions":{"CalendarReq":{"type":"object","properties":{"fieldsComputed":{"type":"integer","format":"int32"},"fieldsNormalized":{"type":"integer","format":"int32"},"firstDayOfWeek":{"type":"integer","format":"int32"},"lenient":{"type":"boolean"},"minimalDaysInFirstWeek":{"type":"integer","format":"int32"},"time":{"type":"string","format":"date-time"},"timeInMillis":{"type":"integer","format":"int64"},"timeZone":{"$ref":"#/definitions/TimeZoneReq"},"weekCountData":{"$ref":"#/definitions/Locale"},"zoneShared":{"type":"boolean"}},"title":"CalendarReq"},"CalendarRes":{"type":"object","properties":{"calendarType":{"type":"string"},"firstDayOfWeek":{"type":"integer","format":"int32"},"lenient":{"type":"boolean"},"minimalDaysInFirstWeek":{"type":"integer","format":"int32"},"time":{"type":"string","format":"date-time"},"timeInMillis":{"type":"integer","format":"int64"},"timeZone":{"$ref":"#/definitions/TimeZoneRes"},"weekDateSupported":{"type":"boolean"},"weekYear":{"type":"integer","format":"int32"},"weeksInWeekYear":{"type":"integer","format":"int32"}},"title":"CalendarRes"},"Electricity":{"type":"object","properties":{"counter":{"type":"integer","format":"int32"},"date":{"$ref":"#/definitions/CalendarRes"},"id":{"type":"integer","format":"int64"},"score":{"type":"number","format":"double"}},"title":"Electricity"},"ElectricityDto":{"type":"object","properties":{"counter":{"type":"integer","format":"int32"},"date":{"$ref":"#/definitions/CalendarReq"}},"title":"ElectricityDto"},"Expense":{"type":"object","properties":{"date":{"$ref":"#/definitions/CalendarRes"},"id":{"type":"integer","format":"int64"},"scoreElect":{"$ref":"#/definitions/Electricity"},"scoreWater":{"$ref":"#/definitions/WaterSupply"}},"title":"Expense"},"ExpenseDto":{"type":"object","properties":{"date":{"$ref":"#/definitions/CalendarReq"},"electId":{"type":"integer","format":"int64"},"waterId":{"type":"integer","format":"int64"}},"title":"ExpenseDto"},"Locale":{"type":"object","title":"Locale"},"ModelAndView":{"type":"object","properties":{"empty":{"type":"boolean"},"model":{"type":"object"},"modelMap":{"type":"object","additionalProperties":{"type":"object"}},"reference":{"type":"boolean"},"status":{"type":"string","enum":["ACCEPTED","ALREADY_REPORTED","BAD_GATEWAY","BAD_REQUEST","BANDWIDTH_LIMIT_EXCEEDED","CHECKPOINT","CONFLICT","CONTINUE","CREATED","DESTINATION_LOCKED","EXPECTATION_FAILED","FAILED_DEPENDENCY","FORBIDDEN","FOUND","GATEWAY_TIMEOUT","GONE","HTTP_VERSION_NOT_SUPPORTED","IM_USED","INSUFFICIENT_SPACE_ON_RESOURCE","INSUFFICIENT_STORAGE","INTERNAL_SERVER_ERROR","I_AM_A_TEAPOT","LENGTH_REQUIRED","LOCKED","LOOP_DETECTED","METHOD_FAILURE","METHOD_NOT_ALLOWED","MOVED_PERMANENTLY","MOVED_TEMPORARILY","MULTIPLE_CHOICES","MULTI_STATUS","NETWORK_AUTHENTICATION_REQUIRED","NON_AUTHORITATIVE_INFORMATION","NOT_ACCEPTABLE","NOT_EXTENDED","NOT_FOUND","NOT_IMPLEMENTED","NOT_MODIFIED","NO_CONTENT","OK","PARTIAL_CONTENT","PAYLOAD_TOO_LARGE","PAYMENT_REQUIRED","PERMANENT_REDIRECT","PRECONDITION_FAILED","PRECONDITION_REQUIRED","PROCESSING","PROXY_AUTHENTICATION_REQUIRED","REQUESTED_RANGE_NOT_SATISFIABLE","REQUEST_ENTITY_TOO_LARGE","REQUEST_HEADER_FIELDS_TOO_LARGE","REQUEST_TIMEOUT","REQUEST_URI_TOO_LONG","RESET_CONTENT","SEE_OTHER","SERVICE_UNAVAILABLE","SWITCHING_PROTOCOLS","TEMPORARY_REDIRECT","TOO_EARLY","TOO_MANY_REQUESTS","UNAUTHORIZED","UNAVAILABLE_FOR_LEGAL_REASONS","UNPROCESSABLE_ENTITY","UNSUPPORTED_MEDIA_TYPE","UPGRADE_REQUIRED","URI_TOO_LONG","USE_PROXY","VARIANT_ALSO_NEGOTIATES"]},"view":{"$ref":"#/definitions/View"},"viewName":{"type":"string"}},"title":"ModelAndView"},"Rate":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"rateE":{"type":"number","format":"double"},"rateWC":{"type":"number","format":"double"},"rateWH":{"type":"number","format":"double"}},"title":"Rate"},"RateDto":{"type":"object","properties":{"rateE":{"type":"number","format":"double"},"rateWC":{"type":"number","format":"double"},"rateWH":{"type":"number","format":"double"}},"title":"RateDto"},"TimeZoneReq":{"type":"object","properties":{"id":{"type":"string"},"rawOffset":{"type":"integer","format":"int32"}},"title":"TimeZoneReq"},"TimeZoneRes":{"type":"object","properties":{"displayName":{"type":"string"},"dstsavings":{"type":"integer","format":"int32"},"id":{"type":"string"},"rawOffset":{"type":"integer","format":"int32"}},"title":"TimeZoneRes"},"View":{"type":"object","properties":{"contentType":{"type":"string"}},"title":"View"},"WaterSupply":{"type":"object","properties":{"counterCold":{"type":"integer","format":"int32"},"counterHot":{"type":"integer","format":"int32"},"date":{"$ref":"#/definitions/CalendarRes"},"id":{"type":"integer","format":"int64"},"scoreCold":{"type":"number","format":"double"},"scoreHot":{"type":"number","format":"double"}},"title":"WaterSupply"},"WaterSupplyDto":{"type":"object","properties":{"counterCold":{"type":"integer","format":"int32"},"counterHot":{"type":"integer","format":"int32"},"date":{"$ref":"#/definitions/CalendarReq"}},"title":"WaterSupplyDto"}}}